Move RemoteImageFeedLoader to production scope
