+++ EssentialDevelopper/EssentialDeveloperFrameworkTests/Feed API/RemoteImageFeedLoaderTests.swift
@@ -9,75 +9,6 @@
 import XCTest
 import EssentialFeed
 
-typealias Closure<T> = (T) -> Void
-typealias VoidClosure = () -> Void
-
-extension HTTPURLResponse {
-    var isOK: Bool { statusCode == 200 }
-}
-
-class HTTPTask: HTTPClientTask {
-    var cancelCompletion: VoidClosure?
-    
-    init(cancelCompletion: @escaping VoidClosure) {
-        self.cancelCompletion = cancelCompletion
-    }
-    
-    func cancel() {
-        cancelCompletion?()
-        cancelCompletion = nil
-    }
-}
-
-class RemoteImageLoadingTask: FeedImageDataLoaderTask {
-    var wrapped: HTTPClientTask?
-    var completion: Closure<FeedImageLoader.ImageLoadResult>?
-    
-    init(completion: @escaping Closure<FeedImageLoader.ImageLoadResult>) {
-        self.completion = completion
-    }
-    
-    func complete(with result: FeedImageLoader.ImageLoadResult) {
-        completion?(result)
-    }
-    
-    func cancel() {
-        completion = nil
-        wrapped?.cancel()
-    }
-}
-
-class RemoteFeedImageLoader: FeedImageLoader {
-    let client: HTTPClient
-    
-    init(client: HTTPClient) {
-        self.client = client
-    }
-    
-    func loadImage(with url: URL, completion: @escaping (ImageLoadResult) -> Void) -> FeedImageDataLoaderTask {
-        let imageLoadTask = RemoteImageLoadingTask(completion: completion)
-        
-        let httpTask = self.client.get(from: url) { [weak self] result in
-            guard let self = self else { return }
-            
-            imageLoadTask.complete(with: result
-                .mapError { _ in ImageLoadingError.connection }
-                .flatMap { (response, data) in
-                    let isValidResponse = response.isOK && !data.isEmpty
-                    return isValidResponse ? .success(data) : .failure(ImageLoadingError.invalidData)
-            })
-        }
-        imageLoadTask.wrapped = httpTask
-        
-        return imageLoadTask
-    }
-}
-
-enum ImageLoadingError: Error {
-    case invalidData
-    case connection
-}
-
 class RemoteImageFeedLoaderTests: XCTestCase {
     
     func test_init_hasNoSideEffects() {
